Integration test to ensure config.cmake overrides hunter.cmake.
This change adds a variable, `HUNTER_PROJECT_ROOT`, to Hunter to allow re-rooting the tree searched for config.cmake and cmake/projects.  If not set, this new variable is set to the same value as `HUNTER_SELF`, but setting it means you can use an alternative set of packages.  The integration test makes use of this to use a dummy test set of packages.

Conflicts:
	.travis.yml