Fix broken MSVC options caused by inserting parallel option.
The `CMAKE_CXX_FLAGS` and `CMAKE_C_FLAGS` options don't need to be saved explicitly to the cache in the toolchain file.  Doing so destroys the default flags that CMake inserts itself.  Instead, append to the _INIT versions of the variables and CMake takes care of including these values, along with its own, once it had read the toolchain file.

Fixes #142.